+++ packages/terra-form-input/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Enabled placeholder support for `fusion-textbox`.
+
 ## 4.22.0 - (April 27, 2023)
 
 * Changed